pyproject.toml
[project]
name = "mac-whisper-speedtest"
version = "0.1.0"
readme = "README.md"
authors = [
    { name = "An Van" },
]

requires-python = ">=3.11"
dependencies = [
    "pyaudio>=0.2.13",
    "faster-whisper>=1.1.1",
    "mlx>=0.5.0",
    "mlx-whisper>=0.4.2",
    "numpy>=2.2.5",
    "rich>=13.7.0",
    "soundfile>=0.12.1",
    "typer>=0.9.0",
    "structlog>=24.1.0",
    "pywhispercpp",
    "insanely-fast-whisper>=0.0.15",
    "hf-xet>=1.0.3",
    "numba>=0.60.0",
    "huggingface-hub>=0.20.0",
    "lightning-whisper-mlx>=0.0.10",
    "parakeet-mlx>=0.3.5",
    "coremltools>=8.0.0",
    "whisper-mps>=0.0.7",
]


[tool.uv.sources]
pywhispercpp = { git = "https://github.com/absadiki/pywhispercpp" }

[project.scripts]
mac-whisper-speedtest = "mac_whisper_speedtest.cli:main"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.1.0",
]
